+++ lang/src/main/scala/com/github/gchudnov/bscript/lang/symbols/ScopeRef.scala
@@ -1,0 +1,6 @@
+package com.github.gchudnov.bscript.lang.symbols
+
+final case class ScopeRef(name: String) extends Scope
+
+object ScopeRef:
+  val global: ScopeRef = ScopeRef("#global")
